<Sprint II> feat: Modelo de Geração de Texto
Correção da duplicidade de dados, geração de respostas em outras línguas e o aprimoramento para reduzir alucinações das respostas geradas, garantido a partir da troca para o modelo GPT2.

+++ files/faiss_search.py
@@ -5,12 +5,12 @@ from .preprocess import preprocess_text
 def criar_faiss_index(embeddings):
     dimension = embeddings.shape[1]
     index = faiss.IndexFlatL2(dimension)
-    index.add(np.array(embeddings))
+    index.add(np.array(embeddings).astype('float32'))  # Certifique-se de que o tipo é float32
     return index
 
 def buscar_por_produto(df, index, consulta, model, k=10):
     consulta_processada = preprocess_text(consulta)  # Pré-processa a consulta
-    consulta_embedding = model.encode([consulta_processada])  # Gera embedding
+    consulta_embedding = model.encode([consulta_processada]).astype('float32')  # Gera embedding e converte para float32
     distances, indices = index.search(np.array(consulta_embedding), k=k)  # Busca no índice
 
     resultados = []
